+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 - `jokeyrhyme-dotfiles env` may now manipulate INFO_PATH and MAN_PATH
 
+- brew: append homebrew to INFO_PATH and MAN_PATH
+
 ### Changed
 
 - env: append homebrew to PATH instead of prepending,